# public/documentation/openapi.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Roamler API",
        "version": "1.0"
    },
    "paths": {
        "/api/product/index": {
            "get": {
                "tags": [
                    "Product"
                ],
                "summary": "List all products",
                "operationId": "064101e99f016542a0667ba29db1c44c",
                "responses": {
                    "200": {
                        "description": "Successful query",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer"
                                                    },
                                                    "name": {
                                                        "type": "string"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Authentication failure"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/product": {
            "put": {
                "tags": [
                    "Product"
                ],
                "summary": "Update a product",
                "operationId": "a37e30aeabe77dfcbcf6f8a412973ea3",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "id": {
                                        "type": "integer",
                                        "example": 123
                                    },
                                    "name": {
                                        "type": "string",
                                        "example": "IPhone 100"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Product is updated"
                    },
                    "400": {
                        "description": "Input data is invalid"
                    },
                    "401": {
                        "description": "Authentication failure"
                    },
                    "404": {
                        "description": "Product not found"
                    },
                    "422": {
                        "description": "Input data is unprocessable"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Product"
                ],
                "summary": "Add a product in database",
                "operationId": "e397660ec66e31cdfa34501a6ceb6beb",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "IPhone 100"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "A product is added"
                    },
                    "400": {
                        "description": "Input data is invalid"
                    },
                    "401": {
                        "description": "Authentication failure"
                    },
                    "422": {
                        "description": "Input data is unprocessable"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/product/{id}": {
            "delete": {
                "tags": [
                    "Product"
                ],
                "summary": "Delete a product",
                "operationId": "9a598c7c634d43c6ce11b243a51f1889",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "The ID of a product",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Product is deleted"
                    },
                    "400": {
                        "description": "Input data is invalid"
                    },
                    "401": {
                        "description": "Authentication failure"
                    },
                    "404": {
                        "description": "Product not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/supermarket/index": {
            "get": {
                "tags": [
                    "Supermarket"
                ],
                "summary": "List all supermarkets",
                "operationId": "b9a8280e888b87bc800976e33d1363d3",
                "responses": {
                    "200": {
                        "description": "Successful query",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer"
                                                    },
                                                    "name": {
                                                        "type": "string"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Authentication failure"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/supermarket": {
            "put": {
                "tags": [
                    "Supermarket"
                ],
                "summary": "Update a supermarket",
                "operationId": "178eecabc9cb3d81e62b999defdf13b6",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "id": {
                                        "type": "integer",
                                        "example": 123
                                    },
                                    "name": {
                                        "type": "string",
                                        "example": "Jumbo"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Supermarket is updated"
                    },
                    "400": {
                        "description": "Input data is invalid"
                    },
                    "401": {
                        "description": "Authentication failure"
                    },
                    "404": {
                        "description": "Supermarket not found"
                    },
                    "422": {
                        "description": "Input data is unprocessable"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Supermarket"
                ],
                "summary": "Add a supermarket in database",
                "operationId": "6a1c8ba330f0ebbff3076990942ebeae",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Albert Heijn"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "A supermarket is added"
                    },
                    "401": {
                        "description": "Authentication failure"
                    },
                    "400": {
                        "description": "Input data is invalid"
                    },
                    "422": {
                        "description": "Input data is unprocessable"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/supermarket/{id}": {
            "delete": {
                "tags": [
                    "Supermarket"
                ],
                "summary": "Delete a supermarket",
                "operationId": "f3f5340886da7eb94868e9849023e3ea",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "The ID of a supermarket",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Supermarket is deleted"
                    },
                    "400": {
                        "description": "Input data is invalid"
                    },
                    "401": {
                        "description": "Authentication failure"
                    },
                    "404": {
                        "description": "Product not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/user/register": {
            "post": {
                "tags": [
                    "User Auth"
                ],
                "summary": "Create a user in the database",
                "operationId": "aaa4aa2bcb8eb835bae113fa4012d935",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "David"
                                    },
                                    "email": {
                                        "type": "string",
                                        "example": "user@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "secretpassword"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "A user is uccessfully registered"
                    },
                    "400": {
                        "description": "Input data is invalid"
                    },
                    "422": {
                        "description": "Input data is unprocessable"
                    }
                }
            }
        },
        "/api/user/login": {
            "post": {
                "tags": [
                    "User Auth"
                ],
                "summary": "User login by email and password",
                "operationId": "eade6bc2477216301854cfd87b8d7488",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "example": "user@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "secretpassword"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful login",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "token": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Invalid email or password"
                    },
                    "400": {
                        "description": "Input data is invalid"
                    },
                    "422": {
                        "description": "Input data is unprocessable"
                    }
                }
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}